created saveTimetableInteractor and moved timetableToFile there
