+++ src/main/java/fileio_use_case/application_business/session_specific_classes/SessionBuilder.java
@@ -22,30 +22,4 @@ public class SessionBuilder {
         }
         return aSession;
     }
-//    /**
-//     * Returns a HashMap<String, Session> class of all sessions (Fall and Winter) based on given HashMap of String
-//     * to CalendarCourse.
-//     * Note: Use .getAllSessions() method in SessionStorer to get
-//     * all Sessions represented as HashMap<String, Session> where the key is the sessionType.
-//     *
-//     * @param allCourses HashMap<String, CalendarCourse>
-//     * @return HashMap<String, Session>
-//     */
-//    public HashMap<String, Session> allSessionBuilder(HashMap<String, CalendarCourse> allCourses)
-//        SessionStorer sessionStore = new SessionStorer();
-//        Session Fall = new Session("F");
-//        sessionStore.addSession("F", Fall);
-//        Session Winter = new Session("S");
-//        sessionStore.addSession("S", Winter);
-//
-//        for (String courseName : allCourses.keySet()) {
-//            CalendarCourse calCourse = allCourses.get(courseName);
-//            if (calCourse.getCourseSession().equals("F")) {
-//                Fall.addCourse(calCourse);
-//            }
-//            else {
-//                Winter.addCourse(calCourse);
-//            }
-//        }
-//        return sessionStore.getAllSessions();
 }
