+++ src/main/java/fileio_use_case/interface_adapters/TimetableFileController.java
@@ -4,7 +4,6 @@ import entities.InvalidSectionsException;
 import fileio_use_case.application_business.FileImportRequestModel;
 import fileio_use_case.application_business.timetable_specific_classes.TimetableFileImportInputBoundary;
 import org.json.simple.parser.ParseException;
-import retrieve_timetable_use_case.application_business.TimetableModel;
 
 import java.io.IOException;
 
@@ -27,14 +26,5 @@ public class TimetableFileController {
     public void createTimetableFile(String filePath) throws IOException, ParseException, java.text.ParseException, InvalidSectionsException {
         FileImportRequestModel requestModel = new FileImportRequestModel(filePath);
         this.interactor.readFromFile(requestModel, "Timetable");
-    }
-
-    /**
-     * Takes in TimetableModel to create Timetable file saved into src/main/saved_timetables/
-     * @param model - a TimetableModel
-     * @param sessionType - the session type of TimetableModel
-     */
-    public void saveTimetable(TimetableModel model, String sessionType) throws InvalidSectionsException {
-        this.interactor.timetableToFile(model, sessionType);
     }
 }
